backend/requirements.txt
Flask
pytesseract
Pillow
google-generativeai
Flask-CORS
